% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{print.varK_calibration}
\alias{print.varK_calibration}
\title{Print method for varK calibration objects}
\usage{
\method{print}{varK_calibration}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"varK_calibration"}.}

\item{digits}{Number of digits for the implied coefficient table.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for varK calibration objects
}
